+++ app/src/main/java/com/example/divvy/Controllers/CreateReviewActivity.java
@@ -8,8 +8,8 @@ import android.widget.EditText;
 import android.widget.RatingBar;
 import android.widget.TextView;
 
+import com.example.divvy.CreateRatingService;
 import com.example.divvy.R;
-import com.example.divvy.RatingService;
 import com.example.divvy.models.Review;
 
 import java.text.SimpleDateFormat;
@@ -58,8 +58,7 @@ public class CreateReviewActivity extends AppCompatActivity{
                     comment_text_field.getText().toString(),
                     username,
                     format.format(new Date()));
-            RatingService.createReviewByUsername(this, review);
-            finish();
+            CreateRatingService.createReviewByUsername(this, review);
         });
 
     }
